Refactor Home page structure: replace HeroSection with ReadyToTransform and remove unused LandingHero and HeroSection components
